backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
google-genai>=1.65.0
google-cloud-firestore>=2.21.0
google-cloud-storage>=2.19.0
google-cloud-texttospeech>=2.24.0
ffmpeg-python>=0.2.0
Pillow>=12.1.0
python-dotenv>=1.1.0
websockets>=16.0
python-multipart>=0.0.20
pydantic>=2.12.0
